+++ ndmg/register.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 """
-ndmg.register.gen_reg
+ndmg.register
 ~~~~~~~~~~~~~~~~~~~~~
 
 Contains ndmg's registration classes, organized as full registration workflows.
@@ -27,7 +27,7 @@ from ndmg.utils import reg_utils
 
 def direct_streamline_norm(streams, fa_path, namer):
     """Applys the Symmetric Diffeomorphic Registration (SyN) Algorithm onto the streamlines to the atlas space defined by .../atlases/reference_brains/FSL_HCP1065_FA_2mm.nii.gz
-    
+
     Parameters
     ----------
     streams : str
@@ -36,7 +36,7 @@ def direct_streamline_norm(streams, fa_path, namer):
         Path to subject's FA tensor image
     namer : NameResource
         variable containing all relevant pathing information
-    
+
     Returns
     -------
     ArraySequence
@@ -110,7 +110,7 @@ def direct_streamline_norm(streams, fa_path, namer):
 
 class DmriReg(object):
     """Class containing relevant paths and class methods for analysing tractography
-    
+
     Parameters
     ----------
     namer : NameResource
@@ -127,7 +127,7 @@ class DmriReg(object):
         skullstrip parameter pre-set. Default is "none"
     simple : bool, optional
         Whether you want to attempt non-linear registration when transforming between mni, t1w, and dwi space. Default is False
-    
+
     Raises
     ------
     ValueError
@@ -317,7 +317,7 @@ class DmriReg(object):
 
     def gen_tissue(self):
         """Extracts the brain from the raw t1w image (as indicated by self.t1w), uses it to create WM, GM, and CSF masks,
-        reslices all 4 files to the target voxel resolution and extracts the white matter edge. Each mask is saved to 
+        reslices all 4 files to the target voxel resolution and extracts the white matter edge. Each mask is saved to
         location indicated by self.map_path
         """
         # BET needed for this, as afni 3dautomask only works on 4d volumes
@@ -511,14 +511,14 @@ class DmriReg(object):
         """alignment from atlas to t1w to dwi. A function to perform atlas alignmet. Tries nonlinear registration first, and if that fails, does a liner
         registration instead.
         Note: for this to work, must first have called t1w2dwi_align.
-        
+
         Parameters
         ----------
         atlas : str
             path to atlas file you want to use
         dsn : bool, optional
             is your space for tractography native-dsn, by default True
-        
+
         Returns
         -------
         str
@@ -693,7 +693,7 @@ class DmriReg(object):
         A function to generate and perform dwi space alignment of avoidance/waypoint masks for tractography.
         First creates ventricle and CC ROI. Then creates transforms from stock MNI template to dwi space.
         NOTE: for this to work, must first have called both t1w2dwi_align and atlas2t1w2dwi_align.
-        
+
         Raises
         ------
         ValueError
@@ -867,7 +867,7 @@ class DmriRegOld:
         self, dwi, gtab, t1w, atlas, aligned_dwi, namer, clean=False, skull="none"
     ):
         """Aligns two images and stores the transform between them
-        
+
         Parameters
         ----------
         dwi : str
@@ -875,7 +875,7 @@ class DmriRegOld:
         gtab : str
             path to file containing gradient driections and strength
         t1w : str
-            path to reference image to be aligned to 
+            path to reference image to be aligned to
         atlas : str
             path to roi atlas file
         aligned_dwi : str
@@ -915,7 +915,7 @@ class DmriRegOld:
 
     def dwi2atlas(self, clean=False):
         """Aligns the dwi image into atlas space
-        
+
         Parameters
         ----------
         clean : bool, optional
