+++ ndmg/scripts/ndmg_dwi_pipeline.py
@@ -13,6 +13,7 @@ For a full description, see here: https://neurodata.io/talks/ndmg.pdf
 import shutil
 import time
 from datetime import datetime
+import os
 
 # package imports
 import nibabel as nib
@@ -20,15 +21,13 @@ from dipy.tracking.streamline import Streamlines
 from subprocess import Popen
 
 # ndmg imports
-import ndmg
 from ndmg import preproc
+from ndmg import register
+from ndmg import track
+from ndmg import graph
 from ndmg.utils import gen_utils
 from ndmg.utils import reg_utils
 from ndmg.utils import cloud_utils
-from ndmg.utils.bids_utils import NameResource
-from ndmg.register import gen_reg
-from ndmg.track import gen_track
-from ndmg.graph import gen_graph
 
 # TODO : not sure why this is here, potentially remove
 os.environ["MPLCONFIGDIR"] = "/tmp/"
@@ -61,7 +60,7 @@ def ndmg_dwi_worker(
     skull="none",
 ):
     """Creates a brain graph from MRI data
-    
+
     Parameters
     ----------
     dwi : str
@@ -112,7 +111,7 @@ def ndmg_dwi_worker(
         Name of the folder on s3 to push to. If empty, push to a folder with ndmg's version number. Default is ""
     skull : str, optional
         skullstrip parameter pre-set. Default is "none".
-    
+
     Raises
     ------
     ValueError
@@ -161,7 +160,7 @@ def ndmg_dwi_worker(
 
     startTime = datetime.now()
 
-    namer = NameResource(dwi, t1w, atlas, outdir)
+    namer = gen_utils.NameResource(dwi, t1w, atlas, outdir)
 
     # TODO : do this with shutil instead of an os command
     print("Output directory: " + outdir)
@@ -353,7 +352,7 @@ def ndmg_dwi_worker(
         print("Running registration in native space...")
 
         # Instantiate registration
-        reg = gen_reg.DmriReg(
+        reg = register.DmriReg(
             namer, nodif_B0, nodif_B0_mask, t1w, vox_size, skull, simple=False
         )
 
@@ -420,14 +419,12 @@ def ndmg_dwi_worker(
 
         # -------- Tensor Fitting and Fiber Tractography ---------------- #
         start_time = time.time()
-        seeds = gen_track.build_seed_list(
-            reg.wm_gm_int_in_dwi, np.eye(4), dens=int(seeds)
-        )
+        seeds = track.build_seed_list(reg.wm_gm_int_in_dwi, np.eye(4), dens=int(seeds))
         print("Using " + str(len(seeds)) + " seeds...")
 
         # Compute direction model and track fiber streamlines
         print("Beginning tractography in native space...")
-        trct = gen_track.RunTrack(
+        trct = track.RunTrack(
             dwi_prep,
             nodif_B0_mask,
             reg.gm_in_dwi,
@@ -473,18 +470,18 @@ def ndmg_dwi_worker(
 
     if reg_style == "native_dsn":
 
-        fa_path = gen_track.tens_mod_fa_est(gtab, dwi_prep, nodif_B0_mask)
+        fa_path = track.tens_mod_fa_est(gtab, dwi_prep, nodif_B0_mask)
 
         # Normalize streamlines
         print("Running DSN...")
-        [streamlines_mni, streams_mni] = gen_reg.direct_streamline_norm(
+        [streamlines_mni, streams_mni] = register.direct_streamline_norm(
             streams, fa_path, namer
         )
 
         # Save streamlines to disk
         print("Saving DSN-registered streamlines: " + streams_mni)
 
-    # TODO : mni space currently broken. Fix EuDX in gen_track.py.
+    # TODO : mni space currently broken. Fix EuDX in track.py.
     # elif reg_style == "mni":
     #     # Check dimensions
     #     start_time = time.time()
@@ -502,7 +499,7 @@ def ndmg_dwi_worker(
 
     #     # Align DWI volumes to Atlas
     #     print("Aligning volumes...")
-    #     reg = gen_reg.DmriRegOld(dwi_prep, gtab, t1w, mask, aligned_dwi, namer, clean)
+    #     reg = register.DmriRegOld(dwi_prep, gtab, t1w, mask, aligned_dwi, namer, clean)
     #     print(
     #         "Registering DWI image at {} to atlas; aligned dwi at {}...".format(
     #             dwi_prep, aligned_dwi
@@ -515,7 +512,7 @@ def ndmg_dwi_worker(
     #     # Compute tensors and track fiber streamlines
     #     print("aligned_dwi: {}".format(aligned_dwi))
     #     print("gtab: {}".format(gtab))
-    #     [tens, streamlines, align_dwi_mask] = gen_track.eudx_basic(
+    #     [tens, streamlines, align_dwi_mask] = track.eudx_basic(
     #         aligned_dwi, gtab, stop_val=0.2
     #     )
     #     tensors = "{}/tensors.nii.gz".format(namer.dirs["output"]["tensor"])
@@ -550,7 +547,7 @@ def ndmg_dwi_worker(
             # align atlas to t1w to dwi
             print("%s%s" % ("Applying native-space alignment to ", labels[idx]))
             labels_im = nib.load(labels_im_file_mni_list[idx])
-            g1 = gen_graph.GraphTools(
+            g1 = graph.GraphTools(
                 attr=len(np.unique(np.around(labels_im.get_data()).astype("int16")))
                 - 1,
                 rois=labels_im_file_mni_list[idx],
@@ -564,7 +561,7 @@ def ndmg_dwi_worker(
             # align atlas to t1w to dwi
             print("%s%s" % ("Applying native-space alignment to ", labels[idx]))
             labels_im = nib.load(labels_im_file_dwi_list[idx])
-            g1 = gen_graph.GraphTools(
+            g1 = graph.GraphTools(
                 attr=len(np.unique(np.around(labels_im.get_data()).astype("int16")))
                 - 1,
                 rois=labels_im_file_dwi_list[idx],
@@ -575,14 +572,14 @@ def ndmg_dwi_worker(
             )
             g1.g = g1.make_graph()
 
-        # TODO : mni-space currently broken. Fix in gen_track.
+        # TODO : mni-space currently broken. Fix in track.
         # elif reg_style == "mni":
         #     labels_im_file = gen_utils.reorient_img(labels[idx], namer)
         #     labels_im_file = gen_utils.match_target_vox_res(
         #         labels_im_file, vox_size, namer, sens="t1w"
         #     )
         #     labels_im = nib.load(labels_im_file)
-        #     g1 = gen_graph.graph_tools(
+        #     g1 = graph.graph_tools(
         #         attr=len(np.unique(np.around(labels_im.get_data()).astype("int16")))
         #         - 1,
         #         rois=labels_im_file,
@@ -604,7 +601,9 @@ def ndmg_dwi_worker(
     # TODO : putting this block of code here for now because it wouldn't run in `ndmg_bids`. Figure out how to put it somewhere else.
     if push and buck and remo is not None:
         if not modif:
-            modif = "ndmg_{}".format(ndmg.__version__.replace(".", "-"))
+            modif = "ndmg_{}".format(
+                __version__.replace(".", "-")
+            )  # TODO : make sure __version__ is in the namespace
         cloud_utils.s3_push_data(buck, remo, outdir, modif, creds, debug=debug)
         print("Pushing Complete!")
         if not debug:
